file update

+++ Merge Sort/Sort.cpp
@@ -83,10 +83,11 @@ int main()
 
     // Print the sorted array
     cout << "Sorted array is: \n";
-    for (int i = 0; i < 10; i++)
+    for (size_t i = 0; i < 10; i++)
     {
         cout << arr[i] << " ";
     }
 
     return 0;
 }
+
